+++ jeap-deploymentlog-docgen/src/main/java/ch/admin/bit/jeap/deploymentlog/docgen/model/ComponentFlowDeploymentDto.java
@@ -9,5 +9,5 @@ public class ComponentFlowDeploymentDto {
     String startedAt;
     String stage;
     String state;
-    String pageId;
+    String pageUrl;
 }